Remove the redirect mapping from "/index.html" to "/"
This is problematic now because the resource handlers have higher
priority than the view controllers (this is intentional, and vital
in fact).

So the "/index.html" mapping in the resource handlers was overriding
the equivalent redirect mapping the view handlers.

But it does not matter, as we can simply have the client side router
perform a client-side redirect from "/index.html" to the application
root instead, thereby keeping the address bar clean.